one more little refactoring
