+++ xchange-bitfinex/src/main/java/org/knowm/xchange/bitfinex/BitfinexExchange.java
@@ -1,4 +1,4 @@
-package org.knowm.xchange.bitfinex.v1;
+package org.knowm.xchange.bitfinex;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -9,7 +9,6 @@ import java.util.stream.Collectors;
 import org.knowm.xchange.BaseExchange;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeSpecification;
-import org.knowm.xchange.bitfinex.BitfinexErrorAdapter;
 import org.knowm.xchange.bitfinex.dto.BitfinexException;
 import org.knowm.xchange.bitfinex.service.BitfinexAccountService;
 import org.knowm.xchange.bitfinex.service.BitfinexAdapters;
