+++ xchange-bitfinex/src/test/java/org/knowm/xchange/bitfinex/v1/service/trade/BitfinexTradeServiceIntegration.java
@@ -11,7 +11,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.ExchangeFactory;
-import org.knowm.xchange.bitfinex.v1.BitfinexExchange;
+import org.knowm.xchange.bitfinex.BitfinexExchange;
 import org.knowm.xchange.bitfinex.v1.service.BitfinexProperties;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.Order;
